@@ -1,11 +1,14 @@
-import { createApp } from "vue";
-import App from "./App.vue";
-createApp(App).mount("#app");
+import './assets/main.css'
 
-import Vue from "vue";
-import router from "./router";
-Vue.config.productionTip = false;
-new Vue({
-  router,
-  render: (h) => h(App),
-}).$mount("#App");
+import { createApp } from 'vue'
+import { createPinia } from 'pinia'
+
+import App from './App.vue'
+import router from './router'
+
+const app = createApp(App)
+
+app.use(createPinia())
+app.use(router)
+
+app.mount('#app')
